pyproject.toml
[tool.poetry]
name = "dyeon-django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.8.12"
Django = "^4.0.6"
django-cors-headers = "^3.13.0"
django-extensions = "^3.2.0"
django-redis = "^5.2.0"
djangorestframework-simplejwt = "^5.2.0"
django-storages = "^1.12.3"
django-heroku = "^0.3.1"
drf-yasg = "^1.21.3"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.20.0"
gunicorn = "^20.1.0"
google-cloud-storage = "^2.4.0"
whitenoise = "^6.2.0"
drf-extra-fields = "^3.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
